+++ cosmo_tester/test_suites/test_blueprints/hello_world_bash_test.py
@@ -22,7 +22,7 @@ from retrying import retry
 
 from cosmo_tester.framework.testenv import TestCase
 from cosmo_tester.framework.git_helper import clone
-from cosmo_tester.framework.assertions import assert_monitoring_data_exists
+from cosmo_tester.framework.test_cases import assert_monitoring_data_exists
 
 
 CLOUDIFY_HELLO_WORLD_EXAMPLE_URL = "https://github.com/cloudify-cosmo/" \
